Added capability to filter by date during a query.
The C score program now operates as either DLL or piped command thanks to #ifdef's.  It can filter by date and threshold.  The python interface is refactored into a class where different methods implement pyscore/cpipe/cdll (previously they were separate functions).

The date filter functions is in the RC config but not exposed on the web interface (that is for a future patch).

First, I must compare to a PubMed query where we query with articles pre-2005, and ask it to retrieve articles post-2005.

The only problem now is that feature scores are generated with counts from ALL of Medline when we should only be using counts from articles in the date range.  Creating the counts requires an additional feature stream pass, doubling the classification time from 30 to 60 seconds.  If I divide the feature stream into years, things will speed up at the expense of greater complexity in the C code.